load plants

+++ plant-project-api/src/main/java/co/simplon/plantproject/dtos/PlantItem.java
@@ -10,8 +10,8 @@ public interface PlantItem {
 
     String getImageUrl();
 
-    Integer getWater();
+    Short getWater();
 
-    Integer getSun();
+    Short getSun();
 
 }
